Add LICENSE file
Philips has kindly allowed the SoftFab code to be open sourced under
a 3-clause BSD license.

The version of SoftFab we are publishing is cleaned up and modernized
(for example, it runs on Python 3 now) after we got the code from
Philips, so I added my and Hans' names to the copyrights as well.