cc: Remove disabling of mark-and-sweep garbage collection
Testing with a Twisted 18.9.0 shows that each request will produce
24 objects that need to be collected. This was 0 with much older
versions of Twisted under Python 2.

It would be useful to dig deeper one day and try to reduce that count,
but even if that would be a success, it would take a while before the
results are deployed. Relying on Twisted not creating reference cycles
was never a good long term strategy.

This does mean we have to move to a real database for the bulk data
sooner, at least before new SoftFab users have created enough task
runs that garbage collection passes become slow.