cc: Replace feed icon by SVG version
The icon was drawn by Matt Brett.
http://www.feedicons.com/
https://mattbrett.com/

We're using it to link an Atom feed, which is an accepted use according
to the guidelines.

I removed the XML and doctype declaration, since those are optional
and increase the file size. I also changed the default presentation
size to 17x17 pixels. I didn't change the actual vectors of the image.

On the Home page, I added "alt" and "title" attributes that were
missing before.