cc: Don't hide user admin pages if authentication is disabled
Running SoftFab without authentication was once a feature for end
users. I don't think this is something we want to encourage, so I'm
turning it into a developer feature instead.

For a developer, it is useful to have the user admin pages enabled
even if there is no need to use them. For example this allows APE
to check a larger percentage of the UI.