cc: Don't call getId() on user objects
Users from the user DB have a getId() method, but predefined users
such as SuperUser do not.

The occurrance in UserDetails was an actual bug, the one in UserList
was impossible to trigger because SuperUser has all privileges and
therefore the second part of the expression would never be evaluated.
But leaving the code as-is would be an accident waiting to happen.