cc: Ask for password only when operator has a name
When authentication is disabled, a nameless operator (instance of
SuperUser) makes the requests. Nameless users cannot have passwords,
so we shouldn't ask for one. Previously we checked whether
authentication was enabled, but I'd like to make that a local
setting of the authentication code rather than a global one.

There can be other nameless users as well (currently UnknownUser
and in the future anonymous guests), but those don't have operator
privileges. Authentication and authorization is split properly,
so there is no risk in not asking for a password for users that
don't have one.