cc: Set SameSite policy on session cookie
This was added to Twisted a few months ago, so we have to check
whether the installed version supports it.

I set the policy to "lax", since "strict" would cause usability
problems: links that take the user to their SoftFab would ignore
their session.

This is another measure that makes CSRF attacks harder.