cc: Use current working directory as database directory
We cannot pass arguments when started from 'twist', so the options
were using the working dir or using an environment variable. The
advantage of using the working dir is that the user doesn't have
to remember the name of an environment variable.

This also catches problems with non-existing or unreadable
directories early.

When using containers, the working directory should be inside the
container anyway, so we might as well use it for something.