cc: Start shadow run cleanup explicitly
Previously, importing the module asked the reactor to call the cleanup
start function with a delay of 0 seconds, but that could lead to it
being called before the database preloading is done. In that case,
the fast path for iterating through the database will yield None
for all records.

Besides fixing the issue described above, this also removes another
import side effect.