cc: Remove default mail domain feature
This was intended to save typing when using SoftFab within a single
organisation. I'm not sure the effort saved is worth the complexity.

For the long term, I'd like to give users the ability to subscribe
to notifications rather than having the ability to add arbitrary
e-mail addresses on configurations and jobs. In that approach,
there is no place for a default mail domain anyway.