cc: Set default dbDir to 'run'
Setting an existing DB directory allows importing modules from an
interactive Python shell. The 'run' dir is used by the "inv run"
developer task, so it has a reasonable chance to exist.

This is necessary because we currently have modules that load data
on import; in the long term we should eliminate such import side
effects.