cc: Remove checkDirect() hack
This worked around our reverse proxy setup not supporting streaming
updates. However, the workaround relies on a particular deployment
configuration that depends on the client being able to connect to
Twisted directly, which may not always be the case.

Since the streaming updates were only necessary for the Notifier
client which we're not going to include in 3.0, I'm removing this
hack. I might remove the actual streaming updates as well later,
just to minimize the amount of code we have to verify and document
before we can release 3.0.