cc: Pass user instead of request to Page.checkAccess()
None of the current implementations uses the request for anything
other than getting the user object.

In general, I think inspecting the request inside checkAccess()
increases the risk of the permission checks getting out of sync with
the code that accesses the database. I want to move in the opposite
direction, where permission checks and database access are more
integrated instead of less. So perhaps checkAccess() will disappear
in the future.