cc: Add type annotations to selectlib module
I changed getCommonTags() to return {} instead of None when called
with zero items. I think it's correct to say that zero items have
no common tags.

This avoids having to put asserts in callers such as the ConfigTags
page when a caller knows it never calls getCommonTags() with zero
items but mypy does not know that.