cc: Add type annotations to PriorityMixin methods
I changed the name of the main group from None to '/' to simplify
the type annotations.

That change may or may not have fixed a bug: PriorityMixin.__lt__()
can compare the names of two tasks/task groups and comparing strings
to None is no longer allowed in Python 3. However, it is not easy
to tell whether this scenario can be triggered in practice.