cc: Map empty string to default for DateArg
In e79c8ff5 (cc: Use stricter value type for page arguments) I made
what I thought was a harmless change to DateArg.parseValue(), to make
the code conform to the return type I wanted to enforce. However, the
result of that change was that date ranges using form text inputs
could no longer be used for optional date arguments.

Unlike the original code, empty string is only accepted and mapped
to None if None is the default value. So it is a bit stricter than
before, but relaxed enough to be practical, I hope.