tr: Remove TODO about replacing unpacker "add" methods
These methods do not only play a role in typing, but also in bridging
singular to plural names: a tag named <capability> triggers a call to
addCapability() which in turn updates the "capabilities" list.

If the "add" method were removed, we'd either have to rename the list
to "capability", which is a confusing name for a collection, or have
the unpacker compute the plural name, which is a bit too fancy for
my liking.