cc: Use AbstractSet for arguments in 'waiting' module
The typing.Set represents a mutable set, but none of these functions
modify the sets passed to them.

I also replaced a cast by an assert: mypy understands that if
an optional value is not None, it is of the declared type.
And for human readers it is now more clear that we are relying
on the code above to set 'missingOnAll' when a non-zero number
of Task Runners is passed but none have all of the required
capabilities, rather than just casting to please mypy.