cc: Use 'name' for first column on ResTypeIndex
Previously 'presentation' was used, which results in the same name
as the 'name' key for non-reserved types. Since we don't show
reserved resource types in this table, that is good enough.

The problem with 'presentation' as the column key is that this
implementation detail was visible to the user, both in the label
and in the URL. The label is easy to override, but the URL is not.