cc: Add ResourceNew page to choose between TR and other resources
We have a different edit page for Task Runners, so we ask the user
what type of resource they want to create. This could have been
done with multiple linkbar buttons instead, but I expect the number
of pre-defined resource types to grow in the future, in which case
the linkbar approach doesn't scale well.