cc: Remove TaskRunner.__applyData() method
This only copied the ID. If the ID is copied on TR creation, there is
no need to copy it again after each time the record is parsed.

This may break compatibility with old TR records, but there are going
to be incompatible changes in the near future as well, so that is not
a big concern. Old TRs records aren't very useful to preserve anyway,
since the actual runners won't be there anymore.