cc: Do no modify getRunners() result
The local group canRunOn() method would take the set returned by
getRunners() as-is and later could apply '&=' to it, which modifies
the original set. The getRunners() methods return their internal
sets, not copies, so this will modify the original set, which is
likely not intended.

This problem was found using mypy, but I didn't want to bury the
fix in the commit that adds the type annotations.