cc: Let ShowReport handle bad job ID in processing
Previously it was handled in presentation, but while that works for
the main page content, it fail to address widgets. As a result, the
auto-refresh script would be fetching updates for non-existing jobs
and the widget code isn't written to deal with 'proc.job' being None.

Initially I tried to make the widget code robust against a None job,
but this requires a lot of extra code and there isn't always an
obviously correct way to deal with it. Besides, requesting
a non-existing entity should result in an HTTP status in the 4xx
range, not a 200.