cc: Treat new Task Runners as unavailable in reason-for-waiting
New Task Runners were being treated as free, but there is no telling
how long it takes for the agent to appear, so it's actually better
to treat them similar to lost TRs for the purpose of task dispatching.

I renamed StatusLevel.LOST to MISSING and mapped both
ConnectionStatus.LOST and NEW to that level.