cc: Avoid use of @runtime decorator
While I like the mechanism, using it would require a runtime dependecy
on the typing_extensions module, which I want to avoid for now.

Previously xmlgen worked around this by using abstract base classes,
but that required a lot of custom code to make isinstance() work.
Instead, we now use hasattr() for the runtime check and use a type
cast to satisfy mypy.