cc: Add record type annotation for taskDefDB
The actual annotation is a one-line change, but quite a bit needed to
be changed to avoid warnings from mypy about the additional code that
it can now check.

I wanted to express that SelectableRecord is both Selectable and
a DatabaseElem, but a TypeVar can only have one bound type. The
SelectableRecordABC workaround I had previously designed for that
didn't actually work because TaskDef only inherited from the same
bases, not from SelectableRecordABC itself.

To make TaskDef inherit from SelectableRecordABC without getting
multiple inheritance paths to DatabaseElem, I had to make TaskDefBase
inherit from SelectableRecordABC.

As a side effect, FrameworkDef is now also taggable, even though
there is no code yet that uses this feature. For the long term,
I want to merge framework and taskdefs, and the merged defs will
be taggable, so this side effect is in line with future plans.