cc: Remove automatic project target 'unknown'
Previously, if a project defined no targets, an implicit target named
"unknown" was added. While this simplified the code, it went against
the idea of users not having to know about features they're not using.

Now a project can just define 0 targets. The Execute dialog still
inserts the 'unknown' target, but that will be changed soon.