cc: Apply defaults at the start of Config.createJob()
This way the defaults are only applied once, even if we'll be creating
multiple jobs in a future commit.

There is a subtle change in the way the 'comment' argument is handled:
previously if the empty string was passed, the Config's comment was
used instead, while now that only happens if None was passed. I think
the new behavior is more consistent with expectations.