cc: Don't use PresentableError when DialogStep.verify() fails
Instead, use a custom exception for dialogs (VerificationError).

The reason for this is that the way the error is handled is not the
same, so using the same exception type is confusing. Also the message
type is different: PresentableError wraps an XML node that typically
contains an entire paragraph, while VerificationError wraps just the
error message itself, excluding the paragraph tag.