cc: Fix computation of initial step
In "cc: Simplify return type of getInitial() dialog method" an
accidental change in behavior was introduced: the 'path' argument
contains the step already visited, so it shouldn't include the
next step to present to the user.

When jumping to the start of the dialog, 'path' should be empty
(default value), but walkSteps() couldn't handle that yet: it
always started from the first step in 'path', while it should be
starting from the first step in 'steps' page attribute.

When jumping to the Task Runner selection step, we now put both
the target and task steps in 'path'. This results in the user
being able to go back to those steps when using the "Load" link
from the LoadExecute page. I think this is an improvement in
usability, so I didn't restore the old behavior.