cc: Handle rowspan correctly
Support for skipping cells covered by rowspan was never implemented,
but until "cc: Improve error detection and reporting in table rows"
the only way this could show is when a column style would be applied
to the wrong cell, which apparently didn't happen in practice.

With the additional sanity checks though, we have to update the
column index for skipped cells to avoid triggering the checks on
correct tables using rowspan.