cc: Do not present a last sync time for a TR that hasn't synced yet
The __lastSyncTime member contains the time the record object was
initialized for Task Runners that haven't been synced since they
were created or loaded. This is used to declare loaded Task Runners
lost if they are not synced in time. But while this is useful
internally, presenting it as a sync time to the user is confusing,
so don't do that.