cc: Fix empty check of 'values' argument for setTag()
We used to demand a Sequence since we check whether it is empty, but
in ScheduleEdit we pass a generator to it (from code that wasn't type
checked until now). This can lead to a key considered to be present
even though there are no tags set for that key. I don't know if this
would lead to problems in practice or whether it is just inefficient.