cc: Fix handling of absent page arguments in getPageURL()
If the argument parsing fails, 'req.args' is None, while the code in
getPageURL() was expecting it to be unset. This probably broke when
type annotations were added.

A better solution would be to either not have the page arguments in
the request object at all, or to postpone construction of the request
object until after argument parsing. But that is a much too invasive
change to do now.