cc: Fix handling of invalid arguments in ShowTaskInfo
Some of the table implementations cannot handle proc.job or proc.task
being None. This is an issue even if the table isn't presented, since
queries are run before the presentation phase.

Another issue with the old approach is that if an argument was invalid,
that would be communicated to the user with words, but the HTTP status
code would be 200 (OK).