cc: Clean up handling of job/tasks with no owner
The value "(none)" is now only displayed, not submitted; the empty
string is submitted instead.

If non-existing user names are passed for 'owner', they are dropped
via ArgsCorrected instead of ignored.

The sorting of user names is now guaranteed to put the absent user
first. I don't recall whether we allow user names that start with
a character lower than '(', but it's fragile to rely on that even
if it is currently true.

The extra code introduced for this is written in such a way that it
can be re-used for targets.