cc: Modernize Task Runner creation in unit tests
The DataGenerator.createTaskRunner() method is central in this.

The TR version was updated to 3.0.0 and the target was removed from
the synced XML.

Since targets are no longer mandatory, the DataGenerator doesn't have
a default target anymore. Instead, tests that want to test the handling
of targets can explicitly pass arguments for that.