cc: Mark a running task as failed when its TR is removed
Until recently, it wasn't possible to remove a TR unless it was lost,
in which case any running task would already have been marked as
failed. But now that TRs can be removed at any time, we need to handle
the situation in which a running task sees its TR disappear.

It is impossible for the task to finish successfully, since the TR
record removal will invalidate the token that the TR client uses to
report task results. So even if the run itself completes fine, the
CC will reject the TR client's attempt to tell it so.

In addition to a test case for the changed behavior, I added a test
case for marking a task as failed when the TR is lost (this wasn't
tested yet) and a test case for removing a non-TR resource (to
document what the intended behavior is, as well as verifying that
nothing crashes when the task completes).