cc: Fix initialization of proc.parent in TaskEdit_GET
The GET version of the page isn't only used for new tasks, it's also
used to load existing records. In the latter case, we do have to look
up the parent definition.

This got broken during the GET/POST split.

I removed the TODO because the PyLint suppression is already
a reminder that the code is doing something that is not ideal.