cc: Improve code that handles tagKey and tagValue arguments
This should make it easier to apply type annotations to this code.

Most of it is improving readabilit, but there is one subtle change in
behavior: if no tag keys are defined, the old code would always reset
tagKey, while the new code only resets a non-empty tagKey. An empty
tagKey means "show all records" (no filter), which is a valid request
even if no tag keys have been defined.