cc: Store arguments inside notification presenter object
If we have an object anyway, storing the arguments there is a lot
cleaner than passing them to every method.

I also removed the joining of the single line summary: while this
is harmless (the characters of a string will be joined back into
the same string), it is also pointless.