cc: Update finding and instantiating page classes
Require a page class to contain an underscore. This was optional before
the GET/POST split was complete, but now we will reject any page class
that doesn't include an HTTP method, so accepting them when searching
is pointless.

This change allows having a module member that is not a page class but
has the same name as the module.

Provide a detailed error message if calling the page constructor fails.