docs: Remove Roadmap document
I created issues in the GitHub tracker for the ideas in this document
that were still relevant: #9 #10 #11 #12 #13 #14 #15 #16 #17 #18 #19 #20

We want to create a new roadmap at some point, containing visions for
the future rather than concrete feature plans. But when we do that,
it should not be part of the installed documentation, but kept on
for example a web site or wiki instead.