cc: Split documentation Markdown rendering and post-processing
Since all inner nodes of the documentation tree contain tables of
content that include titles and abstracts extracted when rendering
child pages, presentation of the top-level page forced rendering of
the entire documentation tree.

While the current documentation can be rendered quickly, that could
change in the future as we add more content and in particular more
dynamic content.

The solution is to split off the post-processing from the Markdown
rendering. The table of contents needs information that is collected
during the Markdown rendering, but the table is presented during
post-processing. So after the split, only direct child documents
need to be rendered immediately.