doc/wrappers: SF_USER_INPUT_<nb> defined
This variable is used to specify the 'task name' in a product in case this product was provided by the user as input to the task. Thus, this manual input product is not generated by any previous task. It gets a sequence number "<nb>", starting at 0, for every unique user input.