style: Make links dark red, except in tables
The blue is too attention-grabbing, in particular in the documentation,
but also in cross-references on UI pages.

In tables I do want to keep the blue, since the dark red doesn't have
good contrast with some of our status colors. Also in tables it's less
of a problem if the links draw attention, perhaps even good, since the
links are the main way to interact with the tables.