style: Remove 'visited' color for links in tables
The color wasn't easy to distinguish from black. We don't think the
'visited' status is likely to be important to users, so instead of
looking for a different color, we make all links the same color.
There is already a lot of information in tables expressed through
color, after all.