cc: Enable login for documentation pages
The immediate reason for this is that we don't have an authenticator
that uses an active login if it's present, but doesn't require login.
Using NoAuthPage led to the user being shown a login link in the title
bar even if they were already logged in.

In the long term, we have ideas to make the documentation presentation
adapt to the database contents, in particular in the Getting Started
section. If we do that, the documentation pages will contain
potentially sensitive data, which means they shouldn't be publicly
viewable, unless the user enables anonymous guest access.