cc: Initial version of artifacts serving module
The goal is to have a storage pool inside the Control Center, so users
do not have to set up a web server when installing a Factory PC.

Currently the only supported access is reading by logged-in users,
but write access by Task Runners will be added soon.

Long term we'd like to have the ability to run storage servers
independent of the Control Center, but using this same code.

Also in the future this code can act as a basis for a REST-style URL
namespace both for UI pages and for a modernized API.