cc: Allow downloading of artifact as gzip file
If the artifact is requested with the ".gz" extension, we serve it
with a gzip Content-Type and without Content-Encoding, so the user
agent will treat it as a gzip file instead of decompress the data.

This is consistent with how PUT will be implemented. Also it might
be useful for a user who wants to archive artifacts outside a CC.