cc: Add type annotations for ResourceIndex page
One issue was found by mypy: the name of the user who suspends
a resource can be None. This cannot happen in production mode,
since then only named users have permission to suspend resources,
but in developer mode we can have an anonymous operator.

I changed the "is this a Task Runner?" checks to use isinstance()
instead of checking the resource type name, since mypy understands
the former but not the latter.