cc: Do not use 'reserved' key for sorting resources
Instead, use a custom sortKey function in ReservedByColumn.

There is one other remaining user of the 'reserved' key: the
GetResourceInfo API call. A minor bug was squashed here: if a custom
resource was suspended and not reserved, it would report it was
reserved by a user with the empty string as its name. After the
fix, it only returns reservations. Reporting who suspended
a resource is something we might want to add later, when
redesigning the API.