tr: Update Ant project to build against Java 8 runtime
Just passing source level 8 still runs the risk of using class library
features not present in that version. We need to build against the
actual runtime library from Java 8 to prevent that.

Actually we should be building against stubs instead of the library
implementation, to prevent the use of implementation classes.
But that has unfortunately never caught on.

The location of the Java 8 runtime is not standardized nor easy to
query, so I make Ant look for it in a few locations. I checked the
installation locations on Debian and openSUSE Linux; this likely
needs to be extended for other operating systems.