tr: Remove SF_REPORT_URL and SF_PRODUCT_URL wrapper variables
The report URL on the TR is obsolete now that the storage pool
determines the URL.

The product URL is not obsolete yet, but in practice often
a local or network file path is used instead of URLs to pass
along products, so I don't think we need to keep it around.
Soonish we'll have products on storage pools, which does
obsolete TR-side configuration of product URLs.