cc: Properly fix checking for TR-specific user privileges
In "cc: Fix asking SuperUser whether it has TR-specific privileges"
(commit 11c8f628) this issue was fixed for SuperUser, but it exists
for many other user classes as well.

The "tr/*" privilege is once again included in the mapping of
privileges to roles, but not granted to any of the roles.
As a result, only TaskRunnerUser has this privilege (hardcoded),
but lookups of non-existing privileges continue to raise KeyError,
since that might help spotting bugs.