inv: Add task to build the Task Runner
This also installs a copy of the JAR inside the Control Center source
tree, so it can be served by the Control Center web server.

Because Ant is relatively slow to start, the Task Runner is only built
automatically on 'run' if the TR JAR does not exist yet. This means
you have to ask for TR rebuilds explicitly when you're making changes
to the TR, or start the TR directly from its build directory.