cc: Refactor as-is artifact archives to use separate Twisted resource
The code for serving the contents of an archive had very little in
common with the code for serving the archive as-is, so it's clearer
to put these code paths in separate resource classes.

The as-is serving is done by the new PlainArtifact class, which is
subclassed for each archive format. The code that runs on the worker
thread was also moved into this class.