cc: Serve sandboxed artifacts from a unique temporary URL namespace
Some types of artifacts need sandboxing: for example an HTML report
can contain JavaScript and for both security and robustness reasons
we don't want the report's scripts to have access to the Control
Center state.

One of the states a sandboxed artifact is denied access to is the
session cookie. While that is good to protect a user's login from
being abused by malicious scripting in a report, it also means that
the Control Center has no way to decide whether a report is allowed
to load additional resources, since it doesn't know on behalf of
which user the request is made.

To solve this, a temporary URL namespace is created: 'sandbox/<key>',
where the key is a secure random string. This URL exists long enough
for page loading to finish, but not much longer than that: currently
it is set to half a minute. If the sandbox URL has expired, the CC
redirects the request to a non-sandboxed version, which then verifies
the user's permissions and creates a new sandboxed URL. This way,
expired URLs are automatically renewed.

I tried hiding the sandbox key from the browser's location bar using
the "user:pass@" syntax in the sandboxed URL, but modern web browsers
will only accept that syntax directly from the user, not from a site:

  https://github.com/whatwg/fetch/pull/465

A more secure approach would be to serve artifacts from a separate
origin (scheme + hostname + port), but that would significantly
complicate the setup of a Control Center, so I didn't want to require
that. It might be something to implement as an option in the future,
if there are projects with very strict security requirements.