cc: Serve deflate-compressed ZIP entries as gzip streams
Since ZIP and gzip both use deflate compression, we can re-use the
compressed data by adding the right header and footer.

I picked gzip over the 'deflate' encoding: despite the name, 'deflate'
encoding is not raw deflate data but instead contained in a zlib
wrapper. This confusing naming has led to some user agents and proxies
implementing support for it incorrectly. The overhead of gzip is just
a few bytes more, so that seemed worth it to avoid potential issues.