cc: Remove broken check for secure channel
The idea was to refuse login over an insecure (HTTP instead of HTTPS)
channel. However, what we can check is whether Twisted is providing
HTTPS, while the most common setup we use is Apache or NGINX handling
HTTPS and acting as a reverse proxy for the Control Center. In that
case, Twisted reports the channel as not secure, but that "insecure"
channel runs over the loopback interface of the server, while the
connection to the outside world is using HTTPS.