tr: Add pre-built Task Runner JAR to CC resources
I was hoping to avoid this and instead build the Task Runner as part
of the installation process. However, it turns out it's not trivial
to build a JAR that targets an older JRE, since we need to build
with the runtime in the bootclasspath and that requires having the
older JRE installed on the build machine. Not all distributions
ship older JREs, for example Debian includes only one version of
OpenJDK's JRE.