cc: Allow cell content for LinkColumn to be overridden
LinkColumn was only used for action columns like 'Edit' and 'Delete',
which have the action verb as the cell content. But there is no need
to limit it to that and inspired by pageLink() function I split the
construction of the link from the adding of the cell content, so it
is possible to override the latter while keeping the former.