cc: Fix initialization of parameters in TaskEdit_GET
initParamArgs() contains the following comment:
"We rely on validateParamState to put everything in the right order."

However, as part of the GET/POST split, the GET version of the page
no longer called the validate function, under the assumption that
the initialized data would already be valid.

This issue was already fixed for frameworks in f01d0827 (cc: Fix
initialization of parameters in FrameworkEdit_GET), but I failed
to realize that TaskEdit had the same issue.