cc: Adopt 'attrs' library
We have two similar mechanisms at the moment: PageArgs and XMLTag,
neither of which integrates well with Python 3.6 type annotations
for fields.

The 'attrs' library does integrate well with those annotations and
seems to be gaining a lot of traction among developers. So I think
it would be better to implement SoftFab's needs on top of 'attrs'
instead of designing a mechanism from scratch.

The 'attrs' library uses calendar versioning, so use the '>='
operator instead of '^'. The version I picked matches the version
of the 'python3-attr' package in Debian 10.