cc: Add 'secret' parameter to repository
This is used to authenticate calls to webhooks.

I didn't use a password input because those cannot be pre-filled,
which means the user would have to re-enter the secret every time
they edit a repository. Another advantage of this approach is that
copy-paste works, so the secret can easily be transferred to the
callling site without a risk of typos.

This approach isn't as safe as using password inputs, but for this
particular purpose I think it's safe enough.