cc: Fix webhook authentication
If a request for a non-existent or secret-less repo was signed with
the string "dummysecret", this would be considered a valid signature.

This could be abused to test whether a repository with a particular
name exists, which is what the code in question was trying to prevent
by returning only generic error messages to the caller.