cc: Register 'suspended' schedule property as boolProperties
This fixes a bug where schedules would be suspended every time
the CC was restarted.

That bug was introduced a week ago in commit 088d2b33 (cc: Explicitly
support 'bool' as XML attribute value type), in which the text
representation of booleans was changed to lower case and XMLTag was
updated to match. The problem is that this update did not include
properties that weren't registered with XMLTag as boolProperties.