cc: Ignore case on 'done' and 'trigger' schedule properties
This fixes a bug where schedules would be reverted to not-done and
not-triggered every time the CC was restarted.

That bug was introduced a week ago in commit 088d2b33 (cc: Explicitly
support 'bool' as XML attribute value type), in which the text
representation of booleans was changed to lower case and XMLTag was
updated to match. The problem is that this update did not include
properties that weren't registered with XMLTag as boolProperties.

Since the 'done' and 'trigger' properties are each only present for
one particular sequence type, making them boolProperties is not an
option here. Instead, I fixed the custom parse code.