cc: Remove the 'sf.extractor' special parameter
This was used to indicate which frameworks had a separate extraction
wrapper.

We decided to remove support for extraction wrappers, since there is
no longer a strong use case for them. Mid-level data can be extracted
by the execution wrapper, which is easier to implement than a separate
wrapper.

The planned feature that extraction wrappers were designed for was
re-extracting old reports using an updated wrapper to get more/better
mid-level data. However, now that reports are stored as artifacts,
such a feature would be better implemented as a server-side plugin
than via a wrapper.