cc: Split overloaded KeySorter constructor into two class methods
Getting the type annotations to work on the overloaded constructor
was proving too much of a pain. By using class methods, the caller
can explicitly ask for the variant it needs.

The only place where this doesn't work great is in datawidgets,
but the amount of extra code needed is rather small.