cc: Remove partially-implemented requirement tracing feature
This feature was intended to enable linking task definitions for
tests in SoftFab with an external requirements management tool.
While this is something we still want to support in principle,
the number of projects tracing their requirements in such a way
is relatively small. So I only want to resurrect the feature when
we have an actual use case for it and are going to complete it by
connecting to at least one external tool.

I kept the ability to tag frameworks and taskdefs in place, since
that might be useful for other features in the future, such as
creating a configuration that contains every task tagged as "unit
test", similar to how schedules can select configurations based
on their tags.