cc: Have Task Runner selection always enabled at the project level
This mechanism is so useful for testing wrappers and debugging issues
with particular factory PCs that it should always be available.
There is still a checkbox that has to be checked to enable it per
configuration, so this change doesn't clutter the UI much.

In the future the selection mechanism should be expanded to cover all
resource types, not just Task Runners.