tr: Log actual arguments passed to Runtime.exec()
Since 40354546 (tr: It is more efficient to pass the arguments) we
no longer pass the command to run as a single string. However, we
still constructed and logged the single string. This meant that there
was actually no efficiency gain, plus the information logged was less
accurate.

Note that the escaping that was done before only dealt with spaces in
arguments but didn't handle quotes in arguments, so in addition to
shorter code we also squashed a bug.