tr: Ask Ruby wrappers to output UTF-8
This is implemented by setting the default external encoding to UTF-8.

I'm not very familiar with Ruby, so I don't know if setting the
internal encoding as well would be a good idea or would break scripts
that aren't prepared for it.

The encoding of the script itself is UTF-8 by default since Ruby 2.0,
so I don't think it's necessary to force the encoding there.