cc: When an artifact resource is not found, try outside the sandbox
For example the raw gzip/ZIP file download isn't available in the
sandbox.

It is easier to redirect outside of the sandbox than to handle all
possible artifact requests in both locations. Also it is safer to
only use the sandbox access mechanism where it is necessary, since
it is protected by a temporary key rather than a login.