cc: Make Pygments a mandatory dependency
The only reason this was an optional dependency is that it was easy
to make it optional when it was only used for the documentation.
Now that we want to use it in more places, having it as an optional
dependency means extra complications to the code and test process,
which is not worth it, in my opinion.