cc: Remove used-before-assignment suppressions before type checking guards
Issue 3127 is partially fixed in PyLint 2.4.1 and will be fully fixed
in the next minor release, so there is no need to keep the suppressions
around long term. I think it makes more sense to remove all of them
at once than only the ones that were obsoleted by 2.4.1, even if that
means we'll see a few false positives in the short term. We're nowhere
near 0 issues anyway.