cc: Don't remove old mid-level data if new data is added
Previously any new mid-level data for a task run would replace
existing data for that same run. This fit the re-extraction model,
where an improved extraction wrapper would deliver the full set of
data for a task run. Recently we scrapped the re-extraction model,
so this requirement isn't there anymore.

The motivation to change the behavior is that now that we parse test
report on the Control Center, we want to be able to combine mid-level
data from that parsing with the mid-level data from the execution
wrapper.