cc: Add base class for reports and implement new extraction methods
The base class allows us to add more report formats in the future.

The Control Center can use the extraction methods can be used to get
task results directly from the report instead of relying on the
wrapper to extract them. This should make it less work to write
wrappers for tools that use standard report formats.