cc: Drop use of 'eq' argument of attr.s
This argument was suggested by the attrs documentation as
an alternative for the deprecated 'cmp' argument. However,
attrs version 18.2.0, the version in Debian 10, doesn't
have the 'eq' argument yet.

In practice, we don't call any comparison operators on the
ArtifactSandbox class; the removal of comparison operators
was only there as a precaution. So the most practical solution
is to remove the argument.