cc: Include SoftFab CSS on all generated pages
This fixes the suppressing of underline on line numbers in text
artifacts. While the parsing for syntax highlighting is done using
Pygments, we have our own formatter which uses SoftFab styles as
well.

Perhaps text files, when served as an artifact, should be served
in raw form instead. However, that is a separate decision: if an
artifact is formatted as HTML for whatever reason, that HTML should
reference all the CSS it needs to be displayed properly.